gc: Initial iteration towards a safepoint mechanism that allows us to save extra roots when establishing the safepoint.
As a first pass towards online-GC without disrupting existing connections,
safepoints for GC will move to be two phased. In the first phase, establishing
the safepoint will have a chance to pass along more refs which must make it to
the GC'd store. In the final phase, the safepoint will have the opportunity to
release any blocking locks, or finalize in any way, after the given chunks have
already made it to the new store.

This commit adds the double-safepoint callback within ValueStore. It leaves the
behavior in dolt_gc as it was previously, such that the first callback is a
no-op and the second callback tears down all existing connections and
invalidates the existing connection so that it cannot be used.